backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Limiter==3.9.2
Flask-Cors==5.0.0
PyMySQL==1.1.0
Werkzeug==3.1.3
python-dotenv==1.0.1
cryptography==43.0.3
mysql-connector-python==9.1.0
PyJWT==2.8.0
flask-sqlacodegen==2.0.0
mysqlclient==2.2.6
APScheduler==3.11.0
